LayoutTests/platform/ios-sim/fast/regions/cssom/get-regions-by-content-horiz-tb-expected.txt

Test for 90759: [CSSRegions] Rename NamedFlow::getRegionsByContentNode to NamedFlow::getRegionsByContent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regionListArticle  is non-null.
PASS regionListArticle.length is 2
PASS regionListArticle.item(0).id is "region"
PASS regionListArticle.item(1).id is "region2"
PASS regionListSpan is non-null.
PASS regionListSpan.length is 1
PASS regionListSpan.item(0).id is "region"
PASS regionListSpan2 is non-null.
FAIL regionListSpan2.length should be 2. Was 1.
PASS regionListSpan3 is non-null.
PASS regionListSpan3.length is 1
PASS regionListSpan3.item(0).id is "region2"
PASS regionListTextNode is non-null.
PASS regionListTextNode.length is 1
PASS regionListTextNode.item(0).id is "region2"
PASS regionListTextNode2 is non-null.
PASS regionListTextNode2.length is 1
PASS regionListTextNode2.item(0).id is "region2"
PASS regionListTextNode3 is non-null.
PASS regionListTextNode3.length is 1
PASS regionListTextNode3.item(0).id is "region2"
FAIL regionListTextNode2.length should be 2. Was 1.
PASS regionListTextNode2.item(0).id is "region2"
FAIL regionListTextNode2.item(1).id should be region3. Threw exception TypeError: null is not an object (evaluating 'regionListTextNode2.item(1).id')
PASS regionListTextNode32.length is 1
PASS regionListTextNode32.item(0).id is "region3"
PASS regionListP1 is non-null.
PASS regionListP1.length is 1
PASS regionListP1.item(0).id is "region4"
PASS regionListP2 is non-null.
PASS regionListP2.length is 1
PASS regionListP2.item(0).id is "region4"
PASS regionListP3 is non-null.
PASS regionListP3.length is 1
PASS regionListP3.item(0).id is "region4"
PASS regionListP22.length is 1
PASS regionListP22.item(0).id is "region5"
PASS successfullyParsed is true

TEST COMPLETE
